--- applications/satellite/quantities.owl
@@ -1,153 +1,0 @@
-<?xml version="1.0"?>
-<rdf:RDF xmlns="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#"
-     xml:base="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4"
-     xmlns:owl="http://www.w3.org/2002/07/owl#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-     xmlns:xml="http://www.w3.org/XML/1998/namespace"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:untitled-ontology-4="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#">
-    <owl:Ontology rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4"/>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Object Properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs -->
-
-    <owl:ObjectProperty rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs">
-        <rdfs:domain rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-        <rdfs:range rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    </owl:ObjectProperty>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#mi -->
-
-    <owl:ObjectProperty rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#mi">
-        <rdfs:subPropertyOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs"/>
-        <rdfs:domain rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation"/>
-        <rdfs:range rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass"/>
-    </owl:ObjectProperty>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Data properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#unit -->
-
-    <owl:DatatypeProperty rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#unit">
-        <rdfs:domain rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-        <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#string"/>
-    </owl:DatatypeProperty>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass">
-        <owl:equivalentClass>
-            <owl:Restriction>
-                <owl:onProperty rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#unit"/>
-                <owl:hasValue rdf:datatype="http://www.w3.org/2001/XMLSchema#string">kg</owl:hasValue>
-            </owl:Restriction>
-        </owl:equivalentClass>
-        <rdfs:subClassOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-        <rdfs:subClassOf>
-            <owl:Restriction>
-                <owl:onProperty rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#unit"/>
-                <owl:hasValue rdf:datatype="http://www.w3.org/2001/XMLSchema#string">kg</owl:hasValue>
-            </owl:Restriction>
-        </rdfs:subClassOf>
-    </owl:Class>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation">
-        <rdfs:subClassOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-    </owl:Class>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Individuals
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#eq1 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#eq1">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation"/>
-        <mi rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1"/>
-    </owl:NamedIndividual>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass"/>
-    </owl:NamedIndividual>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m2 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m2">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass"/>
-    </owl:NamedIndividual>
-</rdf:RDF>
-
-
-
-<!-- Generated by the OWL API (version 4.5.9.2019-02-01T07:24:44Z) https://github.com/owlcs/owlapi -->
-
