--- applications/satellite/quantities_inferred.owl
@@ -1,120 +1,0 @@
-<?xml version="1.0"?>
-<rdf:RDF xmlns="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-5#"
-     xml:base="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-5"
-     xmlns:owl="http://www.w3.org/2002/07/owl#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-     xmlns:xml="http://www.w3.org/XML/1998/namespace"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:untitled-ontology-4="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#">
-    <owl:Ontology rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-5"/>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Object Properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs -->
-
-    <owl:ObjectProperty rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs"/>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#mi -->
-
-    <owl:ObjectProperty rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#mi">
-        <rdfs:subPropertyOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#inputs"/>
-    </owl:ObjectProperty>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass">
-        <rdfs:subClassOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    </owl:Class>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation -->
-
-    <owl:Class rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation">
-        <rdfs:subClassOf rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-    </owl:Class>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Individuals
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#eq1 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#eq1">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#ExplicitEquation"/>
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#RocketEquation"/>
-        <untitled-ontology-4:inputs rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1"/>
-        <untitled-ontology-4:mi rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1"/>
-    </owl:NamedIndividual>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m1">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass"/>
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    </owl:NamedIndividual>
-    
-
-
-    <!-- http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m2 -->
-
-    <owl:NamedIndividual rdf:about="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#m2">
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Mass"/>
-        <rdf:type rdf:resource="http://www.semanticweb.org/johan/ontologies/2022/5/untitled-ontology-4#Quantity"/>
-    </owl:NamedIndividual>
-</rdf:RDF>
-
-
-
-<!-- Generated by the OWL API (version 4.5.9.2019-02-01T07:24:44Z) https://github.com/owlcs/owlapi -->
-
